account for idle p reclaiming
When a syscall is executed, the M might reclaim an idle P when
returning. Effectively, that can mean that the G gets moved from one
P to another. Hence, we need to mark the G as having been executed
with the delay and pick up on the G if we come back from a syscall
with an idle P.

Change-Id: I86e8bb9e93589b7c580b6709e6941845af116949